pyproject.toml
[project]
name = "boe-resumer"
version = "0.1.0"
description = "PDF summarization pipeline using Gemini API with Telegram delivery"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pypdf>=5.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "google-genai>=1.58.0",
    "pytest>=9.0.2",
    "python-telegram-bot[job-queue]>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
